feat(api): refactor study and output
